+++ packages/components/chip/src/ChipClearButton.styles.tsx
@@ -55,8 +55,12 @@ export const chipClearButtonWrapperStyles = cva(
   }
 )
 
+// eslint-disable-next-line tailwindcss/no-custom-classname
 export const chipClearButtonStyles = cva(
-  ['rounded-full', 'focus-visible:outline-none focus-visible:u-ring-inset'],
+  [
+    'rounded-full p-[--sz-2] [font-size:--sz-8] border-sm',
+    'focus-visible:outline-2 focus-visible:outline focus-visible:outline-offset-2  focus-visible:outline-[blue]',
+  ],
   {
     variants: {
       disabled: {